feat: Implement walk-in patients, fix API error handling, and reschedule functionality
- Backend: Add WalkInPatient model and CRUD endpoints
- Backend: Fix GetAppointments to support both Patient and WalkInPatient
- Backend: Add reschedule appointment endpoint for reception
- Frontend: Enhance API error handling with safe JSON parsing
- Frontend: Fix reschedule button event listener issues
- Frontend: Add auto-logout on 401 unauthorized
- Mobile: Fix network errors and API integration
- Docs: Add SQL migration scripts for walk-in patients
- Docs: Update Sprint 2 and Sprint 3 documentation

+++ Backend/HealthySystem.API/Controllers/ServicesController.cs
@@ -78,7 +78,7 @@ namespace HealthySystem.API.Controllers
 
         // GET: api/services/{id} - Lấy thông tin chi tiết 1 dịch vụ
         [HttpGet("{id}")]
-        public async Task<ActionResult<object>> GetService(int id)
+        public async Task<ActionResult<object>> GetService(long id)
         {
             try
             {
